shared Module

+++ product-ms/src/main/java/tn/esprit/productms/repositories/ProductRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.productms.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.productms.entities.Product;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
